hello quangupdate

+++ app/Http/Controllers/Admin/User.php
@@ -29,11 +29,6 @@ class User extends Controller
         return view('admin.user.index');
     }
 
-    /**
-     * Show the form for creating a new resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
     public function search(Request $request)
     {
         $keyword = $request->keyword;
@@ -44,67 +39,40 @@ class User extends Controller
         }
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function new_user(Request $request)
     {
         return view('admin.user.new_user');
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
-    public function show($id)
-    {
-        //
-    }
-
-    /**
-     * Show the form for editing the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function edit($id)
     {
         $data = Users::findOrFail($id);
         return view('admin.user.edit', compact('data'));
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function updata(Request $request, $id)
     {
-        $user = Users::find($id);
-        $user->role_user = $request->role_user;
-        $user->guilty_user = $request->guilty_user;
-        $user->save();
-        return redirect()->action('Admin_UserController@index');
+        if ($id) {
+            $user = Users::find($id);
+            if ($user == true) {
+                $user->role_user = $request->role_user;
+                $user->guilty_user = $request->guilty_user;
+                $user->save();
+                return redirect()->action('Admin_UserController@index');
+            }
+        }
+        abort(404);
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function delete($id)
     {
-        $user = Users::find($id);
-        $user->delete();
-        return redirect()->action('Admin_UserController@index');
+        if ($id) {
+            $user = Users::find($id);
+            if ($user == true) {
+                $user->delete();
+                return redirect()->action('Admin_UserController@index');
+            }
+        }
     }
     public function delete_all(Request $request)
     {
